last commit before getting rid of software serial
removing software serial to be able to use pcint interrupts for catching
pin changes on the vibration switch. They cant be used together. Now
going to have to use the main Serial port which means the arduino can't
be powered from the usual usb cable and will be powered from the pins
instead so the serial port doesn't interfere.